pi: cast uc_snprintf args to int32_t for ia16 compatibility
uc_snprintf %d reads int32_t via va_arg, but on ia16 int is 16-bit.
Passing bare int pushes 2 bytes; va_arg reads 4, picking up stack
garbage as the upper half.  Cast explicitly to int32_t.

+++ src/user/pi/pi.c
@@ -88,7 +88,7 @@ int save_file(const char *path) {
   E.dirty = 0;
 
   char msg[80];
-  uc_snprintf(msg, sizeof(msg), "Written %d bytes", written);
+  uc_snprintf(msg, sizeof(msg), "Written %d bytes", (int32_t)written);
   ui_set_status(msg);
   return 0;
 }
